message

+++ controller/Backend.php
@@ -9,17 +9,14 @@ class Backend
     
     public function adminIndex()
     {
-        if(!isset($_SESSION['admin']))
-        {
-            throw new Exception("erreur 403");
-        }else{
-            $manager = new PostManager();
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
+       
+        $manager = new PostManager();
         $commtManag = new CommentManager();
         $allPosts = $manager->getAllPosts();
         $allComments = $commtManag->getAllComments();
         $flagComment = $commtManag->getFlagComment();
         require(VIEW.'backend/adminView.php');
-        }
         
     }
     public function login($username, $password)
@@ -50,7 +47,7 @@ class Backend
     }
 
     public function deletePost() {
-
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
         $manager = new PostManager();
         $manager->delete('posts', $_GET['id']);
          header('Location: index.php?action=adminIndex');
@@ -59,6 +56,7 @@ class Backend
     
     public function edit()
     {
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
         $manager = new PostManager();
         $post = $manager->getPost($_GET['id']);
         require(VIEW.'backend/updateView.php');
@@ -66,6 +64,7 @@ class Backend
     }
     public function update($title, $content, $postId)
     {
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
         $manager = new PostManager();
         $update = $manager->updatePost($title, $content, $postId);
         header('Location: index.php?action=adminIndex');
@@ -73,11 +72,13 @@ class Backend
     
     public function creatPost()
     {
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
         require(VIEW.'backend/creatView.php');
     }
 
     public function insertPost($title, $content)
     {
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
         $postManager = new PostManager();
         $createdPost = $postManager->creatPost($title, $content);
          if ($createdPost === true) {
@@ -91,8 +92,16 @@ class Backend
             exit;
         }
     }
+    public function annulflag($commentId)
+    {
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
+        $commentManager = new CommentManager();
+        $flag = $commentManager->annulflag($commentId);
+        header('Location: index.php?action=adminIndex');
 
+    }
     public function deleteComment(){
+        if(!isset($_SESSION['admin'])) throw new Exception("erreur 403");
         $manager = new CommentManager();
         $manager->delete('comments', $_GET['id']);
